requirements.txt
rembg>=2.0.0
Pillow>=9.0.0 